Huge refactor
This is the groundwork for thunderbird support.
