+++ src/compose.ts
@@ -1,2 +1,8 @@
 
-throw new Error("Compose script is empty");
+document.body.innerHTML += "Hello world! " + document.body.id;
+
+const canvas = document.createElement("canvas");
+canvas.id = "canvas";
+canvas.oncontextmenu = () => false;
+document.body.appendChild(canvas);
+
